refactor: move the track list logic to custom hooks
LP-25
